chunked file loader added, defaults to 256K locations, loads all to multiple textures, can load files sequentially
I've added a chunked file loading, and changed the location rendering
step to a sequence of 256K locations per drawCall, in the hope that the
whole file can be loaded in order to get parsed to finally load it all
up to multiple textures. Further, you can load files one after the
other. No unload or deactivation options yet or you'd have to reload the
page. Maybe a few million points go well but then you'd have to
sacrifice a cow or do shifty things to load really large files.