added link to reference dataset by @christlet
I've approached Christophe Leterrier on Twitter about his new looking
work. He briefly introduced me to SMLM:

, one thing to keep in mind is that these are not "images" in the strict
sense: these are merely a plot of coordinates determined by processing a
large number of images. At each X,Y localization, a normalized gaussian
is plotted, with width sigma (corresponding to the precision with which
the single fluorophore was localized, which in turns depend on the
number of photons emitted during the blinking event).

I had the confidence that i can render a bunch of points from a texture
in WebGL. I've pirated that trick in the beginning of the WebGL days
from Joshua Koo, also known as @BlurSpline or some say Graphics Noob.
Maybe i wouldn't say so. +++ https://github.com/zz85 +++

So this happened:
https://twitter.com/christlet/status/791642533059293184